205_GraphQL/requirements.txt
prefect==2.19.4
duckdb==1.1.3
pyarrow==17.0.0
pandas==2.2.2
requests==2.32.3
httpx==0.27.0
jinja2==3.1.4
datamodel-code-generator==0.25.4
pydantic==2.8.2
dbt-core==1.8.2
dbt-duckdb==1.8.2
sqlalchemy==2.0.32
python-dotenv==1.0.1

# REST API 用
fastapi==0.110.0
uvicorn>=0.14.0,<0.29.0

# DB client
psycopg2-binary==2.9.9
pymongo==4.8.0

# GraphQL 用
strawberry-graphql==0.226.0
